Import all past linux-based projects.
This is a mixture of different projects all brought together under a common
cmake build infrastructure.  The README.md has more information but here's
a pertinent extract:
    * cmake: Toolchains and modules for cmake
    * docs: Documentation for the project
    * driver: The Kernel driver for the Whitebox radio board
    * gnuradio: Tools for using the Whitebox as a GNURadio peripheral
    * hdl: Hardware description in MyHDL for the FPGAs DSP flow
    * lib: A userspace library for interacting with the radio board
    * linux-cortexm: Tools for building the uClinux Kernel for the ARM Cortex-M
      class of processors.
    * util: Utility Python scripts
    * bootstrap.sh: A one line installer.

Right now, the hardware description and the library both have tests.  They are
run automatically on `make hdl` in the case of the hardware description, and via
various test_* binaries generated by the library on the target device.